ci: align workflow, release manifest check, and clippy -D clean
- Workflow: add build-chain to repository_dispatch; use only the skip_tests
  dispatch input (match blvm-protocol / gates); disk guard show-df: false; run
  cargo metadata after preparing the crates.io release Cargo.toml; use
  skip_tests for verify/docs/attestation on workflow_dispatch; paired fuzz
  gates unchanged
- Clippy: fix all -D warnings (incl. uninlined format args, FromStr for
  EvictionStrategy, clamp/strip_prefix, merged getdata block branch,
  mem::drop on health-check spawn, remove duplicate inner #![cfg] where the
  parent mod already cfg-gates the file)
- Fuzz: patch blvm-primitives in fuzz/Cargo.toml for local sibling resolution
- Rustfmt: apply across touched sources (incl. network/startup Iroh flow)

+++ src/node/parallel_ibd/download.rs
@@ -36,7 +36,7 @@ fn try_load_local_ibd_block(
         return Ok(None);
     }
     let registry = FeatureRegistry::for_protocol(protocol_version);
-    let ts = block.header.timestamp as u64;
+    let ts = block.header.timestamp;
     let segwit_on = registry.is_feature_active("segwit", height, ts);
     let witnesses = match blockstore.get_witness(&expected_hash)? {
         Some(w) => w,
